mpcontribs-api/requirements.txt
apispec==4.6.0
asn1crypto==1.4.0
blinker==1.4
boltons==21.0.0
boto3==1.17.97
css-html-js-minify==2.5.5
dateparser==1.0.0
dnspython==2.1.0
fastnumbers==3.1.0
fdict==0.8.1
filetype==1.0.7
flasgger @ git+https://github.com/flasgger/flasgger#egg=flasgger-0.9.5
flask-compress==1.10.1
flask-cors==3.0.10
flask-log==0.1.0
flask-marshmallow==0.14.0
flask-mongorest @ git+https://github.com/tschaume/flask-mongorest#egg=flask_mongorest-0.3.1
flask-sse==1.0.0
gunicorn[gevent]==20.1.0
json2html==1.3.0
marshmallow-mongoengine==0.30.2
mimerender @ git+https://github.com/tschaume/mimerender@mpcontribs#egg=mimerender-0.6.1
more-itertools==8.8.0
nbformat==5.1.3
notebook==6.4.0
pint==0.17
pymatgen==2022.0.8
pyopenssl==20.0.1
python-snappy==0.6.0
selenium==3.141.0
uncertainties==3.1.5
unflatten==0.1
websocket_client==1.1.0
zstandard==0.15.2
